Move class TextureInformationRepository to new file.
