+++ Dev/Editor/EffekseerCore/Binary/RendererCommonValues.cs
@@ -7,16 +7,6 @@ using Effekseer.Utl;
 
 namespace Effekseer.Binary
 {
-	internal sealed class TextureInformationRepository
-	{
-		public TextureInformation Texture { get; } = new TextureInformation();
-		public TextureInformation Alpha { get; } = new TextureInformation();
-		public TextureInformation UvDistortion { get; } = new TextureInformation();
-		public TextureInformation Blend { get; } = new TextureInformation();
-		public TextureInformation BlendAlpha { get; } = new TextureInformation();
-		public TextureInformation BlendUvDistortion { get; } = new TextureInformation();
-	}
-
 	class RendererCommonValues
 	{
 		private static readonly MaterialSerializer MaterialSerializerInstance = new MaterialSerializer();
